New VoteOnLowestEffort CardView fully functional, and ready for user test
In VoteOnLowestEffort there is now a list of cards. When a card is
selected it will flip and show that it is selected and the user can
proceed with his choice by clicking the fab in the corner. If the user
want another card, they simply select that and the other card will
deselect. Should be fully functional now.